Streamlit Search UI/requirements.txt
streamlit
weaviate-client==4.6.4
pandas
python-dotenv